+++ Habit Management/Core/Extension/Date.swift
@@ -1,0 +1,18 @@
+//
+//  Date.swift
+//  Habit Management
+//
+//  Created by 한수진 on 6/17/25.
+//
+
+import Foundation
+
+extension Date {
+    func toString() -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "MM월\ndd일"
+        dateFormatter.locale = Locale(identifier: "ko_KR")
+        dateFormatter.timeZone = TimeZone(abbreviation: "KST")
+        return dateFormatter.string(from: self)
+    }
+}
